Refactor websocket and rest into separate  modules
