--- graphwalker-cli/src/test/java/org/graphwalker/cli/RestfulServerTest.java
@@ -1,48 +1,0 @@
-package org.graphwalker.cli;
-
-/*
- * #%L
- * GraphWalker Command Line Interface
- * %%
- * Copyright (C) 2005 - 2014 GraphWalker
- * %%
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * #L%
- */
-
-import org.junit.Test;
-
-/**
- * Created by krikar on 10/10/14.
- */
-public class RestfulServerTest extends CLITestRoot {
-
-    /**
-     * TODO: implement test
-     * This test needs to run asyncrounous.
-     */
-    @Test
-    public void online() {/*
-        String args[] = {"online", "-m", "graphml/online/ShoppingCart.graphml", "random(vertex_coverage(100))"};
-        runCommand(args);
-        Assert.assertThat(errMsg, Is.is(""));
-        Assert.assertThat(outMsg, Is.is("Press Control+C to end..." ));
-    */
-    }
-}
