+++ graphwalker-restful/src/main/java/org/graphwalker/restful/Util.java
@@ -1,4 +1,4 @@
-package org.graphwalker.cli;
+package org.graphwalker.restful;
 
 /*
  * #%L
