+++ graphwalker-websocket/src/main/java/org/graphwalker/websocket/WebSocketServer.java
@@ -1,4 +1,4 @@
-package org.graphwalker.cli.service;
+package org.graphwalker.websocket;
 
 /*
  * #%L
