+++ graphwalker-websocket/src/test/java/org/graphwalker/websocket/WebSocketClient.java
@@ -1,4 +1,4 @@
-package org.graphwalker.cli.service;
+package org.graphwalker.websocket;
 
 /*
  * #%L
