+++ graphwalker-websocket/src/test/resources/json/SmallModel.json
