+++ graphwalker-websocket/src/test/resources/org/graphwalker/websocket/WebSocketFlow.graphml
